Move airlock methods to service
